return the bare p2sh signature from wallet connect and configure sighash type
this makes it so we return the bare p2sh siganture (bare as in, no redeem or unlocking script were added to the signature. i.e. the signature is exactly the one walletconnect returned with no modifications).
this also adds one extra configuration parameter to change the sighash type of the transaction. this is to be used later when signing with SIGHASH_SINGLE vs SIGHASH_ALL in the TPU.