fix walletconnect not signing sighash_single
the sighash_types should be sighashTypes for walletconnect to understand it

also the doc comment was updated based on the returned error message by
walletconnect. the returned error message state that sighashTypes field
is a whitelist of possible sighash_types to allow in the psbt signing
process. such thing is not mentioned explicitly in reown's docs (though
i would say reown's docs is the vauge one here, esp since they take
sighashTypes as an array).

reown's docs about sighashTypes (weird because: 1- it's a list, 2- this field already exists in the psbt):
sighashTypes : Integer[] - (Optional) Specifies which part(s) of the transaction the signature commits to. Default is [1].

walletconnect error message about missing sighashTypes (happens when you try to sign using SIGHASH_SINGLE while a missing sighashTypes defaults to [SIGHASH_ALL]):
Sighash type is not allowed. Retry the sign method passing the sighashTypes array of whitelisted types. Sighash type: SIGHASH_SINGLE