create the status for stats struct and fill it from the state machine
also add a couple of columns for new data in stats table. we will be using the same stats table as before with these extra optional columns.
also the events list is sent over but not stored as json like with legacy swaps. will decide later what do to with them (if they are supposed to be stored)