combine other fields from both swap parties
aside from the taker and maker gui and version, we now also grap the maker pubkey from the maker and the taker pubkey from the taker. (i.e. we won't rely on the taker to get the maker's pubkey). on the downside, if the maker has a modded kdf that doesn't send the status, then we won't get the maker's pubkey (which is their right anyway)