fix wasm pack issue
looks like this new 'blocking' feature in testcontainers is the issue. it might be the testcontainers version bump too.
either way, we don't use testcontainers in wasm (testcontainers is a dev dependency used to launch and manage docker containers that are used for end to end tests, and that's not compatible with wasm anyway)