REVIEW CAREFULLY: implement address_from_pubkey for sia
this also unignores test_bob_sells_dsia_for_dutxo which used to fail because this method wasn't implemented for sia (used to fail in save_my_[t|m]aker_swap_event)

after implementing the said method, the test was still fialing. this is because the very first call to save_my_taker_swap_event, the data is not filled yet, which means we try to get the htlc_pub but we find it None and revert back to persistent pub, which isn't a valid ed25519 pubkey and we fail to get the address for such pubkey.

the first call to save_my_taker_swap_event has completely no data anyway. only after apply_event is called over the swap the data actually populates. this is way save_my_taker_swap_event was moved under apply_events. this might break something, so needs a careful review